Transform the sensory data to a large format and perform a PCA

Transform the sensory data to a large format and perform a PCA

perform_senso_mapping(data_profiles, data_products, ...)

Arguments

data_profiles

Tibble. Table with the sensory data.

data_products

Tibble. Table with the info about products.

...

Other parameters of the FactoMineR::PCA() function.

Value

A list of 2 plotly graphs: the PCA ind plot and the PCA var plot

Examples

data("data_profiles_toy")
data("data_products_toy")

res_mapping <- perform_senso_mapping(
  data_profiles = data_profiles_toy,
  data_products = data_products_toy
)

res_mapping$inter_ind_plot # individuals (products) plot

res_mapping$inter_var_plot # variables (sensory attributes) plot